refactor(workspace): extract surface shell helpers
