+++ packages/workspace/src/front/chrome/artifact-surface/SurfaceShell.tsx
@@ -10,12 +10,15 @@ import type { WorkspaceBridge, CommandResult } from "../../bridge/types"
 import type { WorkspaceState, PanelState } from "../../store/types"
 import { WorkbenchLeftPane } from "../workbench-left/WorkbenchLeftPane"
 import { useRegistry, useSurfaceResolverRegistry } from "../../registry"
-import type {
-  SurfaceOpenRequest,
-  SurfacePanelResolution,
-  SurfaceResolverConfig,
-} from "../../../shared/types/surface"
+import type { SurfaceOpenRequest } from "../../../shared/types/surface"
 import { WORKSPACE_OPEN_PATH_SURFACE_KIND } from "../../../shared/types/surface"
+import {
+  findOpenFilePanel,
+  normalizeSurfaceOpenRequest,
+  normalizeWorkbenchPath,
+  surfacePanelId,
+} from "./surfaceShellHelpers"
+export { normalizeSurfaceOpenRequest, resolvePanelForPath } from "./surfaceShellHelpers"
 
 export interface SurfaceShellTab {
   id: string
@@ -85,49 +88,6 @@ export interface SurfaceShellProps {
 }
 
 const COLLAPSED_WIDTH = 40
-
-export function resolvePanelForPath(
-  path: string,
-  registry: { resolve: SurfaceResolverConfig["resolve"] },
-): SurfacePanelResolution | undefined {
-  return registry.resolve({ kind: WORKSPACE_OPEN_PATH_SURFACE_KIND, target: path })
-}
-
-function normalizeWorkbenchPath(path: string): string {
-  const trimmed = path.trim().replace(/\\/g, "/")
-  const noLeadingDot = trimmed.replace(/^\.\//, "")
-  const normalized = noLeadingDot.replace(/\/+/g, "/")
-  // Security: reject path traversal attempts
-  if (normalized.includes("..")) {
-    throw new Error(`Invalid path: path traversal not allowed`)
-  }
-  return normalized
-}
-
-function findOpenFilePanel(api: DockviewApi, path: string) {
-  const panelId = `file:${path}`
-  return api.getPanel(panelId)
-    ?? api.panels.find((panel) =>
-      (panel.params as Record<string, unknown> | undefined)?.path === path
-    )
-}
-
-export function normalizeSurfaceOpenRequest(
-  request: SurfaceOpenRequest,
-): SurfaceOpenRequest {
-  if (request.kind !== WORKSPACE_OPEN_PATH_SURFACE_KIND) return request
-  return {
-    ...request,
-    target: normalizeWorkbenchPath(request.target),
-  }
-}
-
-function surfacePanelId(
-  request: SurfaceOpenRequest,
-  resolved: SurfacePanelResolution,
-): string {
-  return resolved.id ?? `surface:${request.kind}:${request.target}`
-}
 
 let seqCounter = 0
 function ok(): CommandResult {
